requirements.txt
PyQt5
pytesseract
PyMuPDF
indic-transliteration
Pillow